Try to hush clazy warnings about list detachment. (#1489)
I'm puzzled. This is admittedly a hit and miss at 0500 just to get the
build back green.

The loop iterators are all absolutely alreay const refs. (Because I went
through this once before...) My best guess is that it thinks that
the containers themselves (*global_track_list and *wpt_list) are not const.
Could a method be calld that would cause the whole container to detach?

I don't feel like spending an our on cppreference on this right now.
This is a shotgun attempt.