Inmprove doc for options to  duplicate filter. (#1485)
* feat: Add runtime validation for option types

Add a `get_type()` virtual method to the `Option` class hierarchy to
enable runtime type identification.

Use this new method in `Vecs::validate_args` to validate that the
`argtype` specified in `arglist_t` tables matches the actual type of the
`Option` object. This will catch configuration errors at startup, such
as the one that was present in the kalman filter.

This change also includes several fixes to the `Option` class hierarchy
to resolve build errors that surfaced during development, including
breaking a circular dependency between `defs.h` and `option.h`.

* Revert "feat: Add runtime validation for option types"

This reverts commit 6b3661c3f61102450316a102faeeacaf2e2fe3d7.

* More minor doc for duplicate