use Option::isDefaulted method in kalman filter. (#1490)
* use Option::isDefaulted method in kalman filter.
Note that one of the kalman tests attempted to set q_scale_vel to 1.0,
but since this was the default value the actual value used with the
specified cycling profile was 0.1. To match the original reference file
the test was changed to explictly set q_scale_vel to 0.1.  The same results
are obtained by omitting the q_scale_vel option from that invocation.

* catch regernerate script up.

Note that testo has a -d option that alllows a user
to manual copy the generated reference from a test failure.